requirements.txt
requests
shutil
Pillow
mysql-connector-python
geopy
